refactor(opencl): extract SPIR-V support into SRP microcrate
