+++ crates/bitnet-opencl/Cargo.toml
@@ -13,8 +13,8 @@ rust-version.workspace = true
 
 [dependencies]
 bitnet-common = { path = "../bitnet-common", version = "0.2.1-dev" }
+bitnet-spirv = { path = "../bitnet-spirv", version = "0.2.1-dev" }
 log.workspace = true
-tempfile.workspace = true
 thiserror.workspace = true
 serde = { workspace = true }
 serde_json = { workspace = true }
