+++ crates/bitnet-spirv/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "bitnet-spirv"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+authors.workspace = true
+keywords.workspace = true
+categories.workspace = true
+description = "SRP microcrate for SPIR-V compilation, validation, and caching"
+rust-version.workspace = true
+
+[dependencies]
+tempfile.workspace = true
+thiserror.workspace = true
+
+[lints]
+workspace = true
